+++ lib/DAVACL/PrincipalBackend/PrivatePrincipalBackend.php
@@ -59,6 +59,10 @@ class PrivatePrincipalBackend extends AbstractBackend {
         return $this->principalBackend->getAuthTenantByResourceEmail($email, $tenantType ?? TenantType::Resources);
     }
 
+    function provisionUser(string $email, string $firstname = '', string $lastname = '', ?TenantType $tenantType = null): ?AuthTenant {
+        return $this->principalBackend->provisionUser($email, $firstname, $lastname, $tenantType ?? TenantType::User);
+    }
+
     function getAuthTenantByTeamCalendarEmail($email, ?TenantType $tenantType = null): ?AuthTenant {
         return $this->principalBackend->getAuthTenantByTeamCalendarEmail($email, $tenantType ?? TenantType::TeamCalendars);
     }